+++ apps/web/src/components/cms/panels/EditorPanel.vue
@@ -31,6 +31,7 @@ import PostsPanel from "./PostsPanel.vue";
 import PresencePanel from "./PresencePanel.vue";
 import MusicPanel from "./MusicPanel.vue";
 import PlaytimePanel from "./PlaytimePanel.vue";
+import WrappedPanel from "./WrappedPanel.vue";
 import SiteIdentityPanel from "./SiteIdentityPanel.vue";
 
 /**
@@ -54,6 +55,7 @@ const PANEL: Partial<Record<View, Component>> = {
 	presence: PresencePanel,
 	music: MusicPanel,
 	playtime: PlaytimePanel,
+	wrapped: WrappedPanel,
 };
 
 const {